move fragment_param out of amp-viewer-integration
