+++ extensions/amp-viewer-integration/0.1/variable-service.js
@@ -14,9 +14,6 @@
  * limitations under the License.
  */
 
-import {parseQueryString} from '../../../src/url';
-import {userAssert} from '../../../src/log';
-
 /**
  * @typedef {{ancestorOrigin: function(string, string):string, fragmentParam: function(string, string):string}}
  */
@@ -39,29 +36,7 @@ export class AmpViewerIntegrationVariableService {
       ancestorOrigin: (unusedParam, unusedDefaultValue) => {
         return this.getAncestorOrigin_();
       },
-      fragmentParam: (param, defaultValue) => {
-        return this.getFragmentParamData_(param, defaultValue);
-      },
     };
-  }
-
-  /**
-   * Return the FRAGMENT_PARAM from the original location href
-   * @param {*} param
-   * @param {string} defaultValue
-   * @return {string}
-   * @private
-   */
-  getFragmentParamData_(param, defaultValue) {
-    userAssert(
-      param,
-      'The first argument to FRAGMENT_PARAM, the fragment string ' +
-        'param is required'
-    );
-    userAssert(typeof param == 'string', 'param should be a string');
-    const hash = this.ampdoc_.win.location.originalHash;
-    const params = parseQueryString(hash);
-    return params[param] === undefined ? defaultValue : params[param];
   }
 
   /**
